% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.esonar.R
\name{check.esonar}
\alias{check.esonar}
\title{Check \strong{eSonar} Data}
\usage{
\method{check}{esonar}(x, key = TRUE, coordinates = TRUE, time = TRUE, match = TRUE, ...)
}
\description{
Checks an 'esonar' object for errors.
}
