% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.probe.R
\name{check.probe}
\alias{check.probe}
\alias{check.time.probe}
\title{Check \strong{probe} Data}
\usage{
check.probe(x, ...)

\method{check}{time.probe}(x, ...)
}
\description{
Check \code{probe} data object for errors or inconsistencies.
}
\section{Methods (by class)}{
\itemize{
\item \code{time.probe}: Check 'probe' object times for missing entries or errors.
}}

