% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.scsbio.R
\name{check.scsbio}
\alias{check.scsbio}
\title{Check Snow Crab Biological Data}
\usage{
\method{check}{scsbio}(x, ...)
}
\arguments{
\item{x}{Object.}
}
\description{
Functions for validating \code{scsbio} data objects for issues.
}
\examples{
check(read.scsbio(2020))
}
