% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.scsset.R
\name{check.scsset}
\alias{check.scsset}
\title{Check Snow Crab Tow/Set Data}
\usage{
\method{check}{scsset}(x, ...)
}
\arguments{
\item{x}{Object.}
}
\description{
Functions for validating \code{scsset} data objects for issues.
}
\examples{
check(read.scsset(2020))
}
