% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.tow.id.R
\name{check.tow.id}
\alias{check.tow.id}
\title{Check Trawl Tow Identifiers}
\usage{
check.tow.id(x, ...)
}
\description{
Checks tow IDs from snow crab survey set/tow data for errors or inconsistencies.
}
\examples{
x <- read.scsset(2020)
check.tow.number(x)
}
