% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.tow.number.R
\name{check.tow.number}
\alias{check.tow.number}
\title{Check Trawl Tow Numbers}
\usage{
check.tow.number(x, ...)
}
\arguments{
\item{x}{Data object.}
}
\description{
Checks the tow numbers from snow crab survey set/tow data for errors or inconsistencies.
}
\examples{
x <- read.scsset(2020)
check.tow.number(x)
}
