% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gse.R
\name{gse2gsd}
\alias{gse2gsd}
\title{GSE Data Functions}
\usage{
gse2gsd(x, survey, sort = TRUE)
}
\arguments{
\item{x}{List with fields \code{set}, \code{basket}, \code{catch} and \code{detail}, which are data tables 
imported from a GSE Oracle database.}

\item{survey}{Research survey, see \code{\link[gulf.metadata]{project}} for options.}
}
\value{
List with \code{set}, \code{cat} and \code{bio} fields, which are
data frames of set, catch and biological data.
}
\description{
Convert from Groundfish Survey Entry (GSE) format to Groundfish Survey Data (GSD) format.

This function reformats the raw data from a GSE database into a standard
format.  It renames fields, parses variables, converts coordinates to a
decimal-degree format, adds auxiliary fields and homogenizes the index keys
between each table.

The Gulf Survey Data format separates the data into three tables.
Information pertaining to individual sets are stored in the \code{set}
table. Data on individual catches separated by species are stored in the
\code{cat} table.  Data on individual specimens are stored in the \code{bio}
table. Note that there is no length-frequency summary table, though it may
be derived the above tables.
}
\section{Functions}{
\itemize{
\item \code{gse2gsd}: Convert GSE tables to a Gulf survey data object.
}}

\examples{
# Convert imported data to 'gsd' format:
y <- gse2gsd(x)
}
\seealso{
\code{\link[gulf]{read.gse}}, \code{\link[gulf]{gsd2card}}
}
