% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msg.R
\name{msg}
\alias{msg}
\alias{msg.default}
\alias{msg.data.frame}
\title{Return Message Strings}
\usage{
msg(x, ...)

\method{msg}{default}(x, msg = "", echo = FALSE, ...)

\method{msg}{data.frame}(x, msg = "", key, names, echo = FALSE, ...)
}
\arguments{
\item{x}{Object.}

\item{msg}{Character string specifying the message to be returned.}

\item{echo}{Logical value specifying whether to print strings to the R console.}

\item{key}{Character string specifying an index key.}

\item{names}{Character string specifying variable names to be included in the output string.}
}
\value{
Character string of concatenated data fields and messages (invisible).
}
\description{
Functions to build and report message strings.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Default msg method.

\item \code{data.frame}: Generate message from data frame entries.
}}

\examples{
x <- data.frame(year = 2010:2014, site = LETTERS[1:5])
key(x) <- "year"  # Define index key.

# Console output:
msg(x, "Sampling was performed.", echo = TRUE)
msg(x, paste("Record number", 1:nrow(x), "is ok."), echo = TRUE)

# String output:
m <- msg(x, "Sampling was performed.")

}
