% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.scsset.scsbio.R
\name{as.scsset.scsbio}
\alias{as.scsset.scsbio}
\title{Snow Crab Set Data from Biological Data}
\usage{
\method{as.scsset}{scsbio}(x, collapse = FALSE, ...)
}
\arguments{
\item{x}{\code{scsbio} object.}
}
\description{
Functions to compile snow crab set data from snow crab biological data.
}
\section{Methods (by class)}{
\itemize{
\item \code{scsbio}: Compile snow crab set data from snow crab biological data.
}}

\examples{
x <- read.scsbio(1995)
as.scsset(x)
}
