% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gse.R
\name{ese2gsd}
\alias{ese2gsd}
\alias{gse2gsd}
\alias{gsd2oracle}
\title{Data Functions}
\usage{
ese2gsd(x, survey, sort = TRUE)

gse2gsd(x, survey, sort = TRUE)

gsd2oracle(x, survey = "rv")
}
\arguments{
\item{x}{List with fields \code{set}, \code{basket}, \code{catch} and \code{detail}, which are data tables 
imported from a GSE or ESE Oracle database.}

\item{survey}{Research survey, see \code{\link[gulf.metadata]{project}} for options.}
}
\value{
List with \code{set}, \code{cat} and \code{bio} fields, which are data frames of set, catch and biological data.
}
\description{
Functions to convert and reformat data from Groundfish Survey Entry (GSE) and 
             Ecosystrem Survey Entry (GSE) to standardized formats. The Gulf Survey Data format separates the 
             data into three tables. Information pertaining to individual sets are stored in the \code{set}
             table. Data on individual catches separated by species are stored in the \code{cat} table.  Data on 
             individual specimens are stored in the \code{bio} table. Note that there is no length-frequency summary 
             table, though it may be derived the above tables.
}
\section{Functions}{
\itemize{
\item \code{ese2gsd}: Convert GSE (Gulf Survey Entry) or ESE (Ecological Survey Entry) data to a Gulf survey data object.

\item \code{gse2gsd}: Convert GSE tables to a Gulf survey data object.

\item \code{gsd2oracle}: Convert from to GSD to Oracle import format.
}}

\examples{
# Convert imported data to 'gsd' format:
y <- gse2gsd(x)
z <- gsd2oracle(y)
}
